+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api/ChatApi.kt
@@ -229,6 +229,7 @@ internal interface ChatApi {
     fun truncateChannel(
         channelType: String,
         channelId: String,
+        systemMessage: Message?,
     ): Call<Channel>
 
     @CheckResult